refactor: remove or replace adapter.ts
